organizes fcomponents into UI and Expenses
