+++ src/components/Expenses/ExpenseItem.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Card from './Card';
+import Card from '../UI/Card';
 import ExpenseDate from "./ExpenseDate";
 import "./ExpenseItem.css";
