+++ src/components/Expenses/Expenses.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Card from './Card';
+import Card from '../UI/Card';
 import ExpenseItem from './ExpenseItem';
 
 import './Expenses.css'
